IMPROVEMENT: rename tempcal_IMU to tempcal_imu to please pylint
